Continue styling different pages, consolidate Home component into Form
