+++ src/components/Card/Card.css
@@ -1,5 +1,7 @@
 .card {
-  margin: 15px;
+  flex: 1 1 auto;
+  min-width: 300px;
+  margin: auto;
   border: 1px solid #ccc;
   box-shadow: 0 2px 4px rgba(0,0,0,0.1);
   overflow: hidden;
