+++ src/components/EditCaption/EditCaption.css
@@ -1,0 +1,39 @@
+.edit-caption-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+}
+
+.edit-caption-image {
+  max-width: 100%;
+  height: auto;
+  margin-bottom: 20px;
+}
+
+.edit-caption-input {
+  padding: 10px 20px;
+  font-size: 16px;
+  margin-bottom: 20px;
+  border: none;
+  border-radius: 5px;
+  box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+  width: 90%;
+  box-sizing: border-box;
+}
+
+.edit-caption-button {
+  padding: 10px 20px;
+  font-size: 16px;
+  margin: 5px;
+  color: #fff;
+  background-color: #007bff;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: background-color 0.2s;
+}
+
+.edit-caption-button:hover {
+  background-color: #0056b3;
+}
