+++ src/components/EditCaption/EditCaption.jsx
@@ -1,5 +1,7 @@
 import { useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
+import PropTypes from 'prop-types';
+import './EditCaption.css';
 
 function EditCaption({ onSave }) {
   const location = useLocation();
@@ -12,16 +14,24 @@ function EditCaption({ onSave }) {
   };
 
   return (
-    <div>
-      <img src={location.state.image} alt="Caption Edit" />
+    <div className="edit-caption-container">
+      <img src={location.state.image} alt="Caption Edit" className="edit-caption-image" />
       <input
         type="text"
         value={caption}
         onChange={(e) => setCaption(e.target.value)}
+        placeholder="Enter a caption"
+        className="edit-caption-input"
       />
-      <button onClick={handleSave}>Save</button>
+      <button onClick={handleSave} className="edit-caption-button">Save Caption</button>
     </div>
   );
 }
+
+
+EditCaption.propTypes = {
+  onSave: PropTypes.func.isRequired,
+};
+
 
 export default EditCaption;
