+++ src/components/Form/Form.css
@@ -1,5 +1,53 @@
-.success-message {
-  color: green;
-  margin-top: 10px;
-  font-weight: bold;
+.form-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+  width: 100%;
+  max-width: 800px;
+  margin: auto;
+}
+
+.form-image {
+  max-width: 100%;
+  height: auto;
+  margin-bottom: 20px;
+}
+
+.form-input {
+  width: 100%;
+  padding: 10px 20px;
+  margin-bottom: 20px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+}
+
+.form-button {
+  padding: 10px 20px;
+  font-size: 16px;
+  margin: 5px;
+  color: #fff;
+  background-color: #007bff;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: background-color 0.2s;
+}
+
+.form-button:hover {
+  background-color: #0056b3;
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  width: 100%;
+}
+
+.button-group {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
